requirements.txt
gradio==5.8.0
pydantic>=2.0
Pillow
google-generativeai
python-dotenv
google-generativeai
python-dotenv

